MainWeatherData change:
- returns Double instead of BigDecimal

+++ src/main/java/com/darfik/skycast/weatherapi/MainWeatherData.java
@@ -4,17 +4,14 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 
 @Getter
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class MainWeatherData {
 
     @JsonProperty("temp")
-    private BigDecimal temp;
+    private Double temp;
 
     @JsonProperty("feels_like")
-    private BigDecimal feelsLike;
-
+    private Double feelsLike;
 }
